requirements.txt
Flask==3.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
google-cloud-aiplatform[preview]>=1.38.2
vertexai
requests==2.28.1
